moved LS to its own contribution.
