+++ arduino-ide-extension/src/browser/arduino-ide-frontend-module.ts
@@ -307,6 +307,7 @@ import { Daemon } from './contributions/daemon';
 import { FirstStartupInstaller } from './contributions/first-startup-installer';
 import { Notifications } from './contributions/notifications';
 import { OpenSketchFiles } from './contributions/open-sketch-files';
+import { InoLanguage } from './contributions/ino-language';
 
 MonacoThemingService.register({
   id: 'arduino-theme',
@@ -706,6 +707,7 @@ export default new ContainerModule((bind, unbind, isBound, rebind) => {
   Contribution.configure(bind, StartupTasks);
   Contribution.configure(bind, Notifications);
   Contribution.configure(bind, OpenSketchFiles);
+  Contribution.configure(bind, InoLanguage);
 
   // Disabled the quick-pick customization from Theia when multiple formatters are available.
   // Use the default VS Code behavior, and pick the first one. In the IDE2, clang-format has `exclusive` selectors.
